Modularization of state_handling.c file into specific responsibilities.
