+++ src/common/CMakeLists.txt
@@ -19,7 +19,7 @@ target_link_libraries(common
     pico_stdlib          # Base I/O functions
     pico_stdio_usb
     hardware_clocks
-    #hardware_powman #used only for RP2350
+    hardware_powman #used only for RP2350
 )
 
 # Link Wi-Fi driver if supported
